# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/v1/account/me:
    get:
      tags:
      - Api сервиса аккаунта
      summary: Получение данных акаунта
      description: "Получение данных акаунта текущего пользователя. <br> 
        Принцип реализации сервиса Account
        Вам надо реализовать интерфейс контроллера, контроллер, и сам сервис. 
        <br> Общая работа подразумевает:<br>контроллер принима\
        ет в теле запроса параметры и передает их в сервис,<br>Сервис выполняет ло\
        гику и возвращает в контроллер DTO<br>Контроллер оборачивает ответ сервиса\
        \ в ResponseEntity<br>Для работы со временем используйте тип ZonaDateTime<br>д\
        ля работы c boolean используйте Boolean"
      operationId: getMe
      responses:
        "403":
          description: Доступ к данным запрещен.
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Не верный запрос.
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Неизвестная ошибка.
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Метод успешно выполнен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDto'
        "401":
          description: Авторизация не пройдена. Для использования метода необходимо
            авторизоваться.
          content:
            '*/*':
              schema:
                type: string
    put:
      tags:
      - Api сервиса аккаунта
      summary: Обновление данных акаунта
      description: Обновление данных акаунта текущего пользователя
      operationId: putMe
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountDto'
        required: true
      responses:
        "403":
          description: Доступ к данным запрещен.
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Не верный запрос.
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Неизвестная ошибка.
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Метод успешно выполнен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDto'
        "401":
          description: Авторизация не пройдена. Для использования метода необходимо
            авторизоваться.
          content:
            '*/*':
              schema:
                type: string
    delete:
      tags:
      - Api сервиса аккаунта
      summary: Удаление аккаунта
      description: "Удаление аккаунта авторизированного(текущего) пользователя, "
      operationId: deleteMe
      responses:
        "403":
          description: Доступ к данным запрещен.
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Не верный запрос.
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Неизвестная ошибка.
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Метод успешно выполнен.
          content:
            application/json:
              schema:
                type: boolean
        "401":
          description: Авторизация не пройдена. Для использования метода необходимо
            авторизоваться.
          content:
            '*/*':
              schema:
                type: string
...
  /api/v1/post/{id}:
    get:
      tags:
      - Api сервиса постов
      description: Отправка запроса на получение поста по его идентификатору
      operationId: get
      parameters:
      - name: id
        in: path
        description: Идентификатор поста для запроса на получение
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "400":
          description: Не верный запрос.
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Метод успешно выполнен.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostDto'
        "401":
          description: Авторизация не пройдена. Для использования метода необходимо
            авторизоваться.
          content:
            '*/*':
              schema:
                type: string
    delete:
      tags:
      - Api сервиса постов
      description: Отправка запроса на удаление поста
      operationId: deleteById
      parameters:
      - name: id
        in: path
        description: Идентификатор поста для запроса на удаление
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Метод успешно выполнен.
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Неверный запрос.
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Авторизация не пройдена. Для использования метода необходимо
            авторизоваться.
          content:
            '*/*':
              schema:
                type: string
...
  /api/v1/account/{id}:
    get:
      tags:
      - Api сервиса аккаунта
      summary: Получение данных аккаунта по ID
      description: "Получение данных аккаунта по ID тип-UUID, "
      operationId: getId
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "403":
          description: Доступ к данным запрещен.
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Не верный запрос.
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Неизвестная ошибка.
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Метод успешно выполнен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDto'
        "401":
          description: Авторизация не пройдена. Для использования метода необходимо
            авторизоваться.
          content:
            '*/*':
              schema:
                type: string
    delete:
      tags:
      - Api сервиса аккаунта
      summary: Удаление аккаунта по ID
      description: "Удалениеаккаунта по ID тип-UUID, "
      operationId: deleteId
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "403":
          description: Доступ к данным запрещен.
          content:
            '*/*':
              schema:
                type: string
        "400":
          description: Не верный запрос.
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Неизвестная ошибка.
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Метод успешно выполнен.
          content:
            application/json:
              schema:
                type: boolean
        "401":
          description: Авторизация не пройдена. Для использования метода необходимо
            авторизоваться.
          content:
            '*/*':
              schema:
                type: string
  /api/v1/account/search:
    get:
      tags:
      - Api сервиса аккаунта
      summary: Поиск аккаунта(ов) по заданным данным
      description: "Поиск аккаунта(ов) по заданным данным в AccountSearchDto, Pagable\
        \ -  структура данных которую формирует и передает фронт, для нее DTO не н\
        адо делать. Метод должен вернуть объект типа Page<AccountDto><br> Поле showFriends\
        \ нужно для отображения друзей текущего пользователя"
      operationId: getResultSearch
      parameters:
      - name: accountSearchDto
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/AccountSearchDto'
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "403":
          description: Доступ к данным запрещен.
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: Метод успешно выполнен.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        "400":
          description: Не верный запрос.
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Неизвестная ошибка.
          content:
            '*/*':
              schema:
                type: string
        "401":
          description: Авторизация не пройдена. Для использования метода необходимо
            авторизоваться.
          content:
            '*/*':
              schema:
                type: string
components:
  schemas:
    LikeReaction:
      type: object
      properties:
        reactionType:
          type: string
          description: Тип реакции лайка
          example: heart
        count:
          type: integer
          description: Количество реакций в посте для каждого лайка
          format: int32
          example: 1
      description: DTO для работы с реакциями лайков
    PostDto:
      type: object
      properties:
        id:
          type: string
          description: id записи
          format: uuid
        isDeleted:
          type: boolean
          description: Пометка удаления
          default: false
        time:
          type: string
          description: Время публикации поста
          format: date-time
        timeChanged:
          type: string
          description: Время изменения поста
          format: date-time
        authorId:
          type: string
          description: Id автора поста
          format: uuid
        title:
          type: string
          description: Заголовок поста
          default: Про животных
        type:
          type: string
          description: Тип поста (опубликован или запланирован)
          enum:
          - POSTED
          - QUEUED
        postText:
          type: string
          description: "Текст поста, его содержимое"
          default: Содержимое поста
        isBlocked:
          type: boolean
          description: Заблокирован ли автор поста
          default: false
        commentsCount:
          type: integer
          description: Количество комментариев к посту
          format: int32
          default: 2
        reactionType:
          type: array
          description: Список реакций лайков
          items:
            $ref: '#/components/schemas/LikeReaction'
        myReaction:
          type: string
          description: Моя реакция
        likeAmount:
          type: integer
          description: Количество лайков к посту
          format: int32
          default: 1
        myLike:
          type: boolean
          description: Есть ли мой лайк к посту
        imagePath:
          type: string
          description: "Путь к изображению, которое используется в посте. Формируе\
            тся автоматически после добавления изображения в пост"
          default: https://res.cloudinary.com/net/image/upload/v1700829702/fizdftx3pyrpa36vnhpp.jpg
        publishDate:
          type: string
          description: Дата публикации поста
          format: date-time
        tags:
          type: array
          description: Список тегов к посту
          items:
            $ref: '#/components/schemas/TagDto'
      description: "DTO для работы с постами (создание, редактирование)"
    TagDto:
      type: object
      properties:
        id:
          type: string
          description: id записи
          format: uuid
        isDeleted:
          type: boolean
          description: Пометка удаления
          default: false
        name:
          type: string
          description: Название тега
          example: '#полезное'
      description: DTO для работы с тегами
...
    AccountDto:
      type: object
      properties:
        id:
          type: string
          description: id записи
          format: uuid
        isDeleted:
          type: boolean
          description: Пометка удаления
          default: false
        firstName:
          type: string
          description: Имя пользователя
        lastName:
          type: string
          description: Фамилия пользователя
        email:
          type: string
          description: Email
        password:
          type: string
          description: пароль
        phone:
          type: string
          description: Телефон
        photo:
          type: string
          description: Ссылка на фото пользователя
        profileCover:
          type: string
          description: Ссылка на фото для фона аватара
        about:
          type: string
          description: О себе
        city:
          type: string
          description: Город проживания
        country:
          type: string
          description: Страна проживания
        regDate:
          type: string
          description: Время и дата регистрации
          format: date-time
        statusCode:
          type: string
          description: Статус пользователя для друзей
          enum:
          - FRIEND
          - REQUEST_TO
          - REQUEST_FROM
          - BLOCKED
          - DECLINED
          - SUBSCRIBED
          - NONE
          - WATCHING
          - REJECTING
          - RECOMMENDATION
        birthDate:
          type: string
          description: День рождения
          format: date-time
        messagePermission:
          type: string
          description: Сообщение
        lastOnlineTime:
          type: string
          description: Время и дата последней активности пользователя в социальной
            сети
          format: date-time
        isOnline:
          type: boolean
          description: В сети пользователь или нет
        isBlocked:
          type: boolean
          description: Заблокирован ли пользователь
        emojiStatus:
          type: string
        deletionTimestamp:
          type: string
          description: Время удаления пользователя
          format: date-time
        createdDate:
          type: string
          description: Время и дата регистрации в сети пользователя
          format: date-time
        lastModifiedDate:
          type: string
          description: Время и дата изменения личных данных пользователяпользователя
          format: date-time
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        paged:
          type: boolean
        unpaged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    AccountSearchDto:
      type: object
      properties:
        id:
          type: string
          description: id записи
          format: uuid
        isDeleted:
          type: boolean
          description: Пометка удаления
          default: false
        ids:
          type: array
          description: Список id пользователей запрашиваемых в запросе
          items:
            type: string
            description: Список id пользователей запрашиваемых в запросе
            format: uuid
        blockedByIds:
          type: array
          description: Список id заблокированных пользователей
          items:
            type: string
            description: Список id заблокированных пользователей
            format: uuid
        author:
          type: string
          description: Автор поста
        firstName:
          type: string
          description: Имя пользователя
        lastName:
          type: string
          description: Фамилия пользователя
        city:
          type: string
          description: Город проживания
        country:
          type: string
          description: Страна проживания
        email:
          type: string
          description: Email
        isBlocked:
          type: boolean
          description: Заблокирован ли пользователь
        statusCode:
          type: string
          description: Статус пользователя для друзей
          enum:
          - FRIEND
          - REQUEST_TO
          - REQUEST_FROM
          - BLOCKED
          - DECLINED
          - SUBSCRIBED
          - NONE
          - WATCHING
          - REJECTING
          - RECOMMENDATION
        ageFrom:
          type: integer
          description: Начало диапазона возрастов пользователей для поискового запроса
          format: int32
        ageTo:
          type: integer
          description: Окончание диапазона возрастов пользователей для поискового
            запроса
          format: int32
        showFriends:
          type: boolean
          description: поле для отбора только рекомендованных друзей
    Page:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            type: object
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean